+++ CHANGELOG.md
@@ -185,7 +185,7 @@ This is a corrective release for [0.0.14].
 - Initial implementation
 
 [Unreleased]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.16...HEAD
-[0.0.16]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.15...v0.0.15
+[0.0.16]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.15...v0.0.16
 [0.0.15]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.14...v0.0.15
 [0.0.14]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.13...v0.0.14
 [0.0.13]: https://github.com/trailofbits/pypi-attestation-models/compare/v0.0.12...v0.0.13